feat: list activities in Activity Inventory
I implemented the `activity list` command to display all active
activities in the Activity Inventory. Activities are sorted by their
priority and creation date.

I refactored the source code to put the subcommands into their own
subdirectories. I moved the context code to its own package to resolve
a circular dependency error.
