+++ .idea/fileTemplates/internal/Go Application.go
@@ -90,58 +90,8 @@
 // By using the Software, you acknowledge that you have read this Agreement,
 // understand it, and agree to be bound by its terms and conditions.
 
-package cli
+package main
 
-import (
-	"os"
-	"path"
-
-	appcontext "github.com/nakedsoftware/time/internal/context"
-	"github.com/nakedsoftware/time/internal/database"
-	"github.com/spf13/cobra"
-)
-
-var rootCommand = &cobra.Command{
-	Use:     "time",
-	Version: "0.0.1",
-	Short:   "Naked Time is a time management and time tracking application.",
-	Long: `
-Naked Time is a time management and time tracking application that helps you
-to focus on completing important tasks and analyzing how you are using your
-time. Naked Time implements multiple tools to help you to improve your time
-management skills and productivity. Using Naked Time, you can achieve better
-focus in managing and tracking how much time you are working on activities,
-using your time to focus on achieving value and completing important
-activities, and analyzing if you are using your time effectively or can make
-improvements in your estimation and time management skills.
-`,
-	PersistentPreRunE: func(cmd *cobra.Command, args []string) error {
-		return createDatabase(cmd)
-	},
-	RunE: func(cmd *cobra.Command, args []string) error {
-		return cmd.Usage()
-	},
-}
-
-func createDatabase(cmd *cobra.Command) error {
-	homePath, err := os.UserHomeDir()
-	if err != nil {
-		return err
-	}
-
-	dbDir := path.Join(homePath, ".nakedtime")
-	dbPath := path.Join(dbDir, "time.db")
-	if err := os.MkdirAll(dbDir, 0755); err != nil {
-		return err
-	}
-
-	db, err := database.NewDB(dbPath)
-	if err != nil {
-		return err
-	}
-
-	ctx := appcontext.WithDB(cmd.Context(), db)
-	cmd.SetContext(ctx)
-
-	return nil
+func main() {
+	#[[$END$]]#
 }
